// src/models/CoinDataResponse.ts
import { Currencies } from "./Currencies";
import { CurrenciesDate } from "./CurrenciesDate";
import { Localization } from "./Localization";

export type CoinDataResponse = {
    id: string;
    symbol: string;
    name: string;
    asset_platform_id: number | null;
    block_time_in_minutes: number;
    hashing_algorithm: string;
    categories: string[];
    public_notice: string | null;
    localization: Localization;
    description: Localization;
    links: {
        homepage: string[];
        blockchain_site: string[];
        official_forum_url: string[];
        chat_url: string[];
        announcement_url: string[];
        twitter_screen_name: string | null;
        facebook_username: string | null;
        bitcointalk_thread_identifier: string | null;
        telegram_channel_identifier: string | null;
        subreddit_url: string | null;
        repos_url: {
            github: string[];
            bitbucket: [];
        };
    };
    image: {
        thumb: string;
        small: string;
        large: string;
    };
    country_origin: string;
    genesis_date: string;
    sentiment_votes_up_percentage: number;
    sentiment_votes_down_percentage: number;
    market_cap_rank: number;
    coingecko_rank: number;
    coingecko_score: number;
    developer_score: number;
    community_score: number;
    liquidity_score: number;
    public_interest_score: number;
    market_data: {
        current_price: Currencies;
        total_value_locked: null;
        mcap_to_tvl_ratio: null;
        fdv_to_tvl_ratio: null;
        roi: null;
        ath: Currencies;
        ath_change_percentage: Currencies;
        ath_date: CurrenciesDate;
        atl: Currencies;
        atl_change_percentage: Currencies;
        atl_date: CurrenciesDate;
        market_cap: Currencies;
        market_cap_rank: number | null;
        fully_diluted_valuation: Currencies;
        total_volume: Currencies;
        high_24h: Currencies;
        low_24h: Currencies;
        price_change_24h: number;
        price_change_percentage_24h: number;
        price_change_percentage_7d: number;
        price_change_percentage_14d: number;
        price_change_percentage_30d: number;
        price_change_percentage_60d: number;
        price_change_percentage_200d: number;
        price_change_percentage_1y: number;
        market_cap_change_24h: number;
        market_cap_change_percentage_24h: number;
        price_change_24h_in_currency: Currencies;
        price_change_percentage_1h_in_currency: Currencies;
        price_change_percentage_24h_in_currency: Currencies;
        price_change_percentage_7d_in_currency: Currencies;
        price_change_percentage_14d_in_currency: Currencies;
        price_change_percentage_30d_in_currency: Currencies;
        price_change_percentage_60d_in_currency: Currencies;
        price_change_percentage_200d_in_currency: Currencies;
        price_change_percentage_1y_in_currency: Currencies;
        market_cap_change_24h_in_currency: Currencies;
        market_cap_change_percentage_24h_in_currency: Currencies;
        total_supply: number;
        max_supply: number;
        circulating_supply: number;
        last_updated: string;
    };
    public_interest_stats: {
        alexa_rank: number | null;
        bing_matches: number | null;
    };
    last_updated: string;
};